Fixing up bugs with ScanErrs vs ExifErrs
Making changes to tighten code path for the relationship between modtimes and
exif times. Fixed a bug where we might have used the wrong timeval.

Cleaned up some code paths to be more obvious.

Also cleaned up the types of errors too. So a scan error is not mistakenly
counted as an exif error.