Improved sort testing
By refactoring and designing testdir we now have a much more understandable
system for sort testing.

Biggest improvements are that we now no longer have one huge directory for all
testing to operate on but many different onces specialized for special cases.
It allows for closer exercising of code and also easier debugging.

Counts are no longer hardcoded and are dynamic as we populate different types
of files in the directories.

This had us realize we were detecting but not handling duplicates. Now we
remove them.